requirements.txt
# Core Flask dependencies
Flask==3.0.3
Werkzeug==3.0.3

# Image processing
opencv-python==4.10.0.84
numpy==1.24.3
Pillow==10.4.0

# 3D rendering libraries
trimesh==4.4.0
pyrender==0.1.45
pyglet==2.0.10

# Optional for better face landmarks
dlib==19.24.2

# For GLB file support
pygltflib==1.16.1